chore: reorganized code
