+++ NeuralNetwork/activationFunctions.c
@@ -1,0 +1,5 @@
+#include "activationFunctions.h"
+
+double identity(double x) {
+    return x;
+}
