+++ NeuralNetwork/activationFunctions.h
@@ -1,0 +1,6 @@
+#ifndef ACTIVATION_FUNCTIONS_H
+#define ACTIVATION_FUNCTIONS_H
+
+double identity(double x);
+
+#endif
